+++ server/src/domain/smart-info/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './model-config.dto';
