+++ server/src/domain/system-config/dto/system-config-machine-learning.dto.ts
@@ -1,64 +1,6 @@
-import { ApiProperty } from '@nestjs/swagger';
+import { CLIPConfig, ClassificationConfig, RecognitionConfig } from '@app/domain';
 import { Type } from 'class-transformer';
-import { IsBoolean, IsEnum, IsNumber, IsObject, IsOptional, IsString, IsUrl, Max, Min, ValidateIf, ValidateNested } from 'class-validator';
-
-export enum ModelType {
-  IMAGE_CLASSIFICATION = 'image-classification',
-  FACIAL_RECOGNITION = 'facial-recognition',
-  CLIP = 'clip',
-}
-
-export enum CLIPMode {
-  VISION = 'vision',
-  TEXT = 'text',
-}
-
-export class ModelConfig {
-  @IsBoolean()
-  enabled!: boolean;
-
-  @IsString()
-  @ApiProperty({ type: 'string' })
-  modelName!: string;
-
-  @IsEnum(ModelType)
-  @ApiProperty({ type: 'string', enum: ModelType, enumName: 'ModelType' })
-  modelType!: ModelType;
-}
-
-export class ClassificationConfig extends ModelConfig {
-  @IsNumber()
-  @Min(0)
-  @Max(1)
-  @Type(() => Number)
-  @ApiProperty({ type: 'integer' })
-  minScore!: number;
-}
-
-export class CLIPConfig extends ModelConfig {
-  @IsEnum(CLIPMode)
-  @IsOptional()
-  @ApiProperty({ enumName: 'CLIPMode', enum: CLIPMode })
-  mode?: CLIPMode;
-}
-
-
-export class RecognitionConfig extends ModelConfig {
-  @IsNumber()
-  @Min(0)
-  @Max(1)
-  @Type(() => Number)
-  @ApiProperty({ type: 'integer' })
-  minScore!: number;
-
-  @IsNumber()
-  @Min(0)
-  @Max(2)
-  @Type(() => Number)
-  @ApiProperty({ type: 'integer' })
-  maxDistance!: number;
-}
-
+import { IsBoolean, IsObject, IsUrl, ValidateIf, ValidateNested } from 'class-validator';
 
 export class SystemConfigMachineLearningDto {
   @IsBoolean()
